Respond to review comments, clean up
