+++ packages/expo-notifications/ios/EXNotifications/Permissions/PermissionsModule.swift
@@ -29,7 +29,11 @@ public class PermissionsModule: Module {
     }
 
     AsyncFunction("requestPermissionsAsync") { (requestedPermissions: NotificationPermissionRecord, promise: Promise) in
-      EXUserFacingNotificationsPermissionsRequester.setRequestedPermissions(requestedPermissions.toDictionary())
+      let defaultAuthorizationOptions: UNAuthorizationOptions = [.alert, .badge, .sound]
+      requester.authorizationOptions = requestedPermissions.numberOfOptionsRequested() > 0
+        ? requestedPermissions.authorizationOptionValue()
+        : defaultAuthorizationOptions
+
       appContext?
         .permissions?
         .askForPermission(
@@ -39,43 +43,6 @@ public class PermissionsModule: Module {
         )
     }
   }
+
 }
 
-struct NotificationPermissionRecord: Record {
-  @Field
-  var allowAlert: Bool?
-  @Field
-  var allowBadge: Bool?
-  @Field
-  var allowSound: Bool?
-  @Field
-  var allowDisplayInCarPlay: Bool?
-  @Field
-  var allowCriticalAlerts: Bool?
-  @Field
-  var provideAppNotificationSettings: Bool?
-  @Field
-  var allowProvisional: Bool?
-
-  func count() -> Int {
-    return (self.allowAlert != nil ? 1 : 0) +
-    (self.allowBadge != nil ? 1 : 0) +
-    (self.allowSound != nil ? 1 : 0) +
-    (self.allowDisplayInCarPlay != nil ? 1 : 0) +
-    (self.allowCriticalAlerts != nil ? 1 : 0) +
-    (self.provideAppNotificationSettings != nil ? 1 : 0) +
-    (self.allowProvisional != nil ? 1 : 0)
-  }
-
-  func authorizationOptionValue() -> UNAuthorizationOptions {
-    var options: UNAuthorizationOptions = []
-    if self.allowAlert ?? false { options.insert(.alert) }
-    if self.allowBadge ?? false { options.insert(.badge) }
-    if self.allowSound ?? false { options.insert(.sound) }
-    if self.allowDisplayInCarPlay ?? false { options.insert(.carPlay) }
-    if self.allowCriticalAlerts ?? false { options.insert(.criticalAlert) }
-    if self.provideAppNotificationSettings ?? false { options.insert(.providesAppNotificationSettings) }
-    if self.allowProvisional ?? false { options.insert(.provisional) }
-    return options
-  }
-}
